refactor(oidc-authorize): extract check to its own module (#62)
The `oidc-authorize` check is materially larger than every sibling probe in `src/checks.ts` (PKCE helper, dual-shape redactor, normalized origin+pathname compare, half a dozen fail branches). Move it and its module-private helpers (`generatePkcePair`, `redactCodeAndState`, `normalizePath`) into a dedicated `src/oidc-authorize.ts` so `checks.ts` reads at altitude again. The `checks` array imports the exported constant unchanged; every load-bearing behavior — strict origin+pathname compare, dual-shape redactor, doubled-trailing-slash normalization, code+state redaction on every diagnostic path — is preserved. All existing tests continue to pass.

Closes #62
