+++ docs/code-layout.md
@@ -3,7 +3,8 @@
 | File                         | Purpose                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             |
 | ---------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `src/handler.ts`             | Lambda entry. Loads config, resolves DJ credentials (env or Secrets Manager), runs all checks in parallel, publishes per-check metrics to CloudWatch, throws if any check failed (so the Lambda Errors metric also fires).                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          |
-| `src/checks.ts`              | The check definitions, the `signInDj` helper, and the PKCE `generatePkcePair` helper used by `oidc-authorize` (wxyc-canary#60). One exported `checks` array. Adding a new check = one entry.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        |
+| `src/checks.ts`              | The check definitions and the `signInDj` helper. One exported `checks` array. Adding a new check = one entry. `oidc-authorize` was moved to its own module (`src/oidc-authorize.ts`, wxyc-canary#62) — every other check is small enough to inline here.                                                                                                                                                                                                                                                                                                                                                                                                                                                            |
+| `src/oidc-authorize.ts`      | The `oidc-authorize` check (wxyc-canary#60), extracted from `checks.ts` under wxyc-canary#62. Owns the PKCE `generatePkcePair` helper, the URL/JSON dual-shape `redactCodeAndState`, the trailing-slash-tolerant `normalizePath`, and the strict origin+pathname compare. All helpers are module-private — never exported.                                                                                                                                                                                                                                                                                                                                                                                          |
 | `src/enrichment-check.ts`    | Write-canary orchestrator (v1, opt-in). Inserts a sentinel flowsheet row, polls until LML enrichment populates `youtube_music_url`, deletes the row, ends the show. Best-effort cleanup; skips when another DJ is on-air.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           |
 | `src/client.ts`              | HTTP client wrapping `fetch` with a per-request `AbortController` timeout. Deliberately no retry on the surfaces being measured (a flaky retry hides brownouts).                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    |
 | `src/types.ts`               | Codable shapes for `Check`, `CheckResult`, `CheckOutcome`, `CanaryConfig`.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          |
